Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
 int main() {
-    float r;
-    float pie=3.14;
-    float area;
+    int r;
+    double pie=3.14;
+    double area;
     scanf("%f",r);
     area=pie*r*r;
-    printf("Area: %.2lf\n",area);
+    printf("Area: %.2f\n",area);
     return 0;
 }
